+++ src/react/hooks/__tests__/useFragment.test.tsx
@@ -1835,7 +1835,11 @@ describe("useFragment", () => {
 
   describe("return value `complete` property", () => {
     let cache: InMemoryCache, wrapper: React.FunctionComponent;
-    const ItemFragment = gql`
+    const ItemFragment: TypedDocumentNode<{
+      __typename: "Item";
+      id: number;
+      text: string;
+    }> = gql`
       fragment ItemFragment on Item {
         id
         text
@@ -1876,13 +1880,17 @@ describe("useFragment", () => {
     });
 
     test("if only partial data is available, `complete` is `false`", () => {
-      cache.writeFragment({
-        fragment: ItemFragment,
-        data: {
-          __typename: "Item",
-          id: 5,
-        },
-      });
+      {
+        using _ = spyOnConsole("error");
+        cache.writeFragment({
+          fragment: ItemFragment,
+          // @ts-expect-error purposefully omitting text
+          data: {
+            __typename: "Item",
+            id: 5,
+          },
+        });
+      }
 
       const { result } = renderHook(
         () =>